Improve Vagrant packaging & client config
Add a Vagrantfile to be packaged with the box to enable port forwarding & run
the client-config.sh script when a box is brought up.
Run client-config.sh from the Vagrantfile rather than trying to run it from
Chef.
Improve the client-config.sh script to create the config/ directory before
emiting the config file.